requirements.txt
ray[all]==1.2.0
Flask
Flask-Cors
Flask-GraphQl
graphene
graphene-sqlalchemy
SQLAlchemy
shortuuid
aiohttp_cors
aiortc
simplejson
pyinstrument
sh
xvfbwrapper # TODO: MSWIN not compatible
pygame
opencv_python
ffmpeg-python
gym
tensorflow==2.3.1
torch==1.7.1 
torchvision==0.8.2 
matplotlib
gym
colorama
gputil
msgpack
msgpack_numpy
pytest
psycopg2-binary